+++ apeiron/toolkits/discord/toolkit.py
@@ -3,7 +3,9 @@ from typing import Any
 from langchain_core.tools.base import BaseTool, BaseToolkit
 
 from apeiron.tools.discord.add_reaction import DiscordAddReactionTool
+from apeiron.tools.discord.get_emoji import DiscordGetEmojiTool
 from apeiron.tools.discord.get_guild import DiscordGetGuildTool
+from apeiron.tools.discord.get_message import DiscordGetMessageTool
 from apeiron.tools.discord.get_user import DiscordGetUserTool
 from apeiron.tools.discord.list_channels import DiscordListChannelsTool
 from apeiron.tools.discord.list_emojis import DiscordListEmojisTool
@@ -24,13 +26,14 @@ class DiscordToolkit(BaseToolkit):
             List of Discord tools.
         """
         return [
-            DiscordGetUserTool(client=self.client),
             DiscordAddReactionTool(client=self.client),
+            DiscordGetEmojiTool(client=self.client),
+            DiscordGetGuildTool(client=self.client),
+            DiscordGetMessageTool(client=self.client),
+            DiscordGetUserTool(client=self.client),
+            DiscordListChannelsTool(client=self.client),
             DiscordListEmojisTool(client=self.client),
             DiscordListMessagesTool(client=self.client),
-            DiscordListChannelsTool(client=self.client),
             DiscordReplyMessageTool(client=self.client),
             DiscordSendMessageTool(client=self.client),
-            DiscordGetUserTool(client=self.client),
-            DiscordGetGuildTool(client=self.client),
         ]
